fix: use f32 instead of i32 for volume

+++ src/models/shared/voice.rs
@@ -18,7 +18,7 @@ pub struct UserVoiceSettings {
     /// pan object - set the pan of the user
     pub pan: Option<VoicePan>,
     /// integer - set the volume of user (defaults to 100, min 0, max 200)
-    pub volume: Option<i32>,
+    pub volume: Option<f32>,
     /// boolean - set the mute state of the user
     pub mute: Option<bool>,
 }
